CHANGES: 1) added short class black_scholes_exact and norm_cdf - lss_utility.h 2) finished implementation of general black_scholes_equation and black_scholes_equation_cuda solvers - lss_black_scholes_equation_solvers.h, lss_black_scholes_equation_solvers_cuda.h
3) reworked the explicit schemes to use policies even for cuda version

4) started testing blach-sholes pde

TODO:
1) clean up the code in explicit schemes especially
2) further testing for non-cuda version of black-scholes
3) write unit tests for cuda version of black-scholes

+++ unit_tests/lss_one_dim_space_variable_pure_heat_equation_cuda_t.h
@@ -3170,11 +3170,11 @@ void testExplPureHeatEquationDoubleDirichletBCDeviceEuler() {
   // set initial condition:
   expl_solver.set_initial_condition(initial_condition);
   // set thermal diffusivity (C^2 in PDE)
-  expl_solver.set_2_order_coefficient([](float x) { return 1.0; });
+  expl_solver.set_2_order_coefficient([](double x) { return 1.0; });
   // set convection term in PDE
-  expl_solver.set_1_order_coefficient([](float x) { return 0.0; });
+  expl_solver.set_1_order_coefficient([](double x) { return 0.0; });
   // set zero-order term in PDE
-  expl_solver.set_0_order_coefficient([](float x) { return 0.0; });
+  expl_solver.set_0_order_coefficient([](double x) { return 0.0; });
   // get the solution:
   expl_solver.solve(solution);
   // get exact solution:
